+++ frontend/src/modules/employees/pages/EmployeesPage.vue
@@ -1,6 +1,6 @@
 <script setup>
 import { ref, onMounted } from 'vue'
-import { searchEmployees, saveEmployee, uploadEmployeesExcel } from '../../../services/api'
+import { searchEmployees, saveEmployee, uploadEmployeesExcel } from '../services/employees.service'
 
 const filters = ref({ name: '', status: '', region: '' })
 const employees = ref([])
