+++ frontend/src/modules/report/services/report.service.js
@@ -1,0 +1,5 @@
+import api from '../../../services/api'
+
+export const getReport = () => api.get('/report')
+
+
